rename dokumentService til DistribuerDokumentService & fiks kompilering
