+++ service/src/main/kotlin/no/nav/su/se/bakover/service/dokument/JournalførDokumentServiceImpl.kt
@@ -81,5 +81,4 @@ class JournalførDokumentServiceImpl(
                 KunneIkkeJournalføreBrev.KunneIkkeOppretteJournalpost
             }
     }
-
 }
