+++ service/src/main/kotlin/no/nav/su/se/bakover/service/journalføring/JournalføringService.kt
@@ -1,5 +1,17 @@
 package no.nav.su.se.bakover.service.journalføring
 
-interface JournalføringService {
-    fun journalfør()
+import no.nav.su.se.bakover.service.dokument.JournalførDokumentService
+import no.nav.su.se.bakover.service.skatt.JournalførSkattDokumentService
+
+/**
+ * Wrapper klasse som kaller journalfør() for alle klasser
+ */
+class JournalføringService(
+    private val journalførDokumentService: JournalførDokumentService,
+    private val journalførSkattDokumentService: JournalførSkattDokumentService,
+) {
+    fun journalfør() {
+        journalførDokumentService.journalfør()
+        journalførSkattDokumentService.journalfør()
+    }
 }
